backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.4
pydantic-settings==2.3.4
SQLAlchemy==2.0.31
psycopg2-binary==2.9.9
langgraph==0.1.5
langchain-core==0.2.10
langchain-huggingface==0.0.3
python-jose==3.3.0
python-multipart==0.0.9
requests==2.32.3
passlib==1.7.4
python-dotenv==1.0.1
